Refactor NetMRI Resources
Combined the common code in the NetMRI resources into a mixin
class.

Fixed the NetMRI connection parameters to reflect the updated
infoblox-netmri module.
