+++ doc/templates/member.yaml
@@ -8,14 +8,12 @@ resources:
   grid_member:
     type: Infoblox::Grid::Member
     properties:
-      connection: { url: "https://172.22.128.92/wapi/v2.3", username: admin, password: infoblox, sslverify: False }
+      connection: { url: "https://dnsvnf-gm/wapi/v2.3/", username: admin, password: infoblox, sslverify: False }
       name: nios-test.infoblox.local
       model: IB-VM-820
-      LAN1: 121749db-d096-49f5-a483-1209b55890ef
+      LAN1: { get_param: lan1_port }
       licenses: ["vnios", "dns", "enterprise"]
-      gm_ip: 172.16.98.100
+      gm_ip: 172.22.138.160
 outputs:
-  join_token:
-    value: { get_attr: [grid_member, grid_join_token] }
   user_data:
     value: { get_attr: [grid_member, user_data] }
