+++ doc/templates/netmri-resource.yaml
@@ -13,10 +13,10 @@ parameters:
     description: the command to run
     default: "show ver"
 resources:
-  netmri_job:
+  netmri_resource:
     type: Infoblox::NetMRI::ManagedResource
     properties:
-      connection: { url: "netmri", username: admin, password: per1sc0pe1, sslverify: False }
+      connection: { host: "netmri", api_version: 3, username: admin, password: per1sc0pe1, use_ssl: False }
       create_job:
         source: { script: "Ad Hoc Command Batch" }
       #targets: [ { device_id: { get_param: device_id } } ]
