+++ heat_infoblox/resource_utils.py
@@ -37,14 +37,19 @@ def port_schema(port_name, is_required):
 
 CONN_DESCR = {
     constants.NETMRI: {
-        constants.URL: "The URL to the NetMRI API (example: "
-                       "'https://netmri/api/3.0')",
+        constants.API: "NetMRI API",
+        constants.HOST: "The host name or IP address of the netmri.",
+        constants.API_VERSION: 'The API version, for example "3".',
+        constants.SSLVERIFY: "Flag to verify SSL, or path to a certificate.",
+        constants.USE_SSL: "If true, HTTPS will be used",
         constants.USERNAME: 'The username for NetMRI.',
         constants.PASSWORD: 'The password for NetMRI.'
     },
     constants.DDI: {
+        constants.API: "Infoblox WAPI",
         constants.URL: "The URL to the Infoblox WAPI (example: "
                        "'https://infoblox/wapi/v2.3')",
+        constants.SSLVERIFY: "Flag to verify SSL, or path to a certificate.",
         constants.USERNAME: 'The username for Infoblox.',
         constants.PASSWORD: 'The password for Infoblox.'
     }
@@ -52,31 +57,57 @@ CONN_DESCR = {
 
 
 def connection_schema(conn_type):
-    return properties.Schema(
-        properties.Schema.MAP,
-        required=True,
-        schema={
+    if conn_type == constants.NETMRI:
+        conn_schema = {
+            constants.HOST: properties.Schema(
+                properties.Schema.STRING,
+                CONN_DESCR[conn_type][constants.HOST],
+                required=True
+            ),
+            constants.API_VERSION: properties.Schema(
+                properties.Schema.STRING,
+                CONN_DESCR[conn_type][constants.API_VERSION],
+                required=True
+            ),
+            constants.USE_SSL: properties.Schema(
+                properties.Schema.BOOLEAN,
+                CONN_DESCR[conn_type][constants.USE_SSL],
+                default=True
+            )
+        }
+    else:
+        conn_schema = {
             constants.URL: properties.Schema(
                 properties.Schema.STRING,
                 CONN_DESCR[conn_type][constants.URL],
                 required=True
-            ),
-            constants.USERNAME: properties.Schema(
-                properties.Schema.STRING,
-                CONN_DESCR[conn_type][constants.USERNAME],
-                required=True
-            ),
-            constants.PASSWORD: properties.Schema(
-                properties.Schema.STRING,
-                CONN_DESCR[conn_type][constants.PASSWORD],
-                required=True
-            ),
-            constants.SSLVERIFY: properties.Schema(
-                properties.Schema.BOOLEAN,
-                _('If True, the SSL certificate will be validated.'),
-                default=True
             )
         }
+
+    conn_schema.update({
+        constants.USERNAME: properties.Schema(
+            properties.Schema.STRING,
+            CONN_DESCR[conn_type][constants.USERNAME],
+            required=True
+        ),
+        constants.PASSWORD: properties.Schema(
+            properties.Schema.STRING,
+            CONN_DESCR[conn_type][constants.PASSWORD],
+            required=True
+        ),
+        constants.SSLVERIFY: properties.Schema(
+            properties.Schema.BOOLEAN,
+            CONN_DESCR[conn_type][constants.SSLVERIFY],
+            default=True
+        )
+    })
+
+    api_name = CONN_DESCR[conn_type][constants.API]
+    return properties.Schema(
+        properties.Schema.MAP,
+        "The connection details for the %s" % api_name,
+        required=True,
+        schema=conn_schema
     )
